SellerFormViewModel and Department select component

+++ SalesWebMvc/Controllers/DepartmentsController.cs
@@ -7,20 +7,24 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using SalesWebMvc.Data;
 using SalesWebMvc.Models;
+
 namespace SalesWebMvc.Controllers
 {
     public class DepartmentsController : Controller
     {
         private readonly SalesWebMvcContext _context;
+
         public DepartmentsController(SalesWebMvcContext context)
         {
             _context = context;
         }
+
         // GET: Departments
         public async Task<IActionResult> Index()
         {
             return View(await _context.Department.ToListAsync());
         }
+
         // GET: Departments/Details/5
         public async Task<IActionResult> Details(int? id)
         {
@@ -28,22 +32,26 @@ namespace SalesWebMvc.Controllers
             {
                 return NotFound();
             }
+
             var department = await _context.Department
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (department == null)
             {
                 return NotFound();
             }
+
             return View(department);
         }
+
         // GET: Departments/Create
         public IActionResult Create()
         {
             return View();
         }
+
         // POST: Departments/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        // To protect from overposting attacks, enable the specific properties you want to bind to.
+        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Name")] Department department)
@@ -56,6 +64,7 @@ namespace SalesWebMvc.Controllers
             }
             return View(department);
         }
+
         // GET: Departments/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
@@ -63,6 +72,7 @@ namespace SalesWebMvc.Controllers
             {
                 return NotFound();
             }
+
             var department = await _context.Department.FindAsync(id);
             if (department == null)
             {
@@ -70,9 +80,10 @@ namespace SalesWebMvc.Controllers
             }
             return View(department);
         }
+
         // POST: Departments/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        // To protect from overposting attacks, enable the specific properties you want to bind to.
+        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Name")] Department department)
@@ -81,6 +92,7 @@ namespace SalesWebMvc.Controllers
             {
                 return NotFound();
             }
+
             if (ModelState.IsValid)
             {
                 try
@@ -103,6 +115,7 @@ namespace SalesWebMvc.Controllers
             }
             return View(department);
         }
+
         // GET: Departments/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
@@ -110,24 +123,32 @@ namespace SalesWebMvc.Controllers
             {
                 return NotFound();
             }
+
             var department = await _context.Department
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (department == null)
             {
                 return NotFound();
             }
+
             return View(department);
         }
+
         // POST: Departments/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var department = await _context.Department.FindAsync(id);
-            _context.Department.Remove(department);
+            if (department != null)
+            {
+                _context.Department.Remove(department);
+            }
+
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
+
         private bool DepartmentExists(int id)
         {
             return _context.Department.Any(e => e.Id == id);
